Ensure shapeDistTraveled is set (it is used in PresentationServiceImpl.include) and it is a double

+++ onebusaway-transit-data-federation/src/main/java/org/onebusaway/transit_data_federation/impl/transit_graph/TransitGraphDaoImpl.java
@@ -386,7 +386,7 @@ public class TransitGraphDaoImpl implements TransitGraphDao {
   }
 
   @Override
-  public boolean insertStopTime(AgencyAndId tripId, AgencyAndId stopId, int arrivalTime, int departureTime, int shapeDistanceTravelled) {
+  public boolean insertStopTime(AgencyAndId tripId, AgencyAndId stopId, int arrivalTime, int departureTime, double shapeDistanceTravelled) {
     boolean rc = _graph.insertStopTime(tripId, stopId, arrivalTime, departureTime, shapeDistanceTravelled);
     if (rc) {
       TripEntryImpl trip = (TripEntryImpl) getTripEntryForId(tripId);
